💿 feat: Persist positions to the database
* I want to see if there's a median/mean max profit for each symbol. Is this
this the correct take profit?

+++ cpp/sandbox/src/alpaca/tao_bot/set_status.cpp
@@ -46,7 +46,7 @@ void Alpaca::TaoBot::set_status(order_t *order) {
 
   if (original_status != order_status_t::ORDER_FILLED &&
       order->status == order_status_t::ORDER_FILLED) {
-    order->timestamp = (int)std::time(nullptr);
+    order->timestamp = (double)std::time(nullptr);
 
     const char *order_action = Alpaca::constants::ORDER_ACTIONS[order->action];
     const char *log_icon = this->ICONS[order_action];
